Moving WorkflowGraphSectionControls out into own file.
