+++ src/encoded/static/components/item-pages/WorkflowView.js
@@ -7,13 +7,12 @@ import _ from 'underscore';
 import { 
     ItemPageTitle, ItemHeader, ItemDetailList, TabbedView, AuditTabView, AttributionTabView,
     ExternalReferenceLink, FilesInSetTable, FormattedInfoBlock, WorkflowDetailPane,
-    WorkflowNodeElement, CollapsibleItemViewButtonToolbar
+    WorkflowNodeElement, CollapsibleItemViewButtonToolbar, WorkflowGraphSectionControls
 } from './components';
 import DefaultItemView from './DefaultItemView';
 import { console, object, DateUtility, Schemas, isServerSide, navigate, layout } from './../util';
 import Graph, { parseAnalysisSteps, parseBasicIOAnalysisSteps, DEFAULT_PARSING_OPTIONS } from './../viz/Workflow';
 import { requestAnimationFrame } from './../viz/utilities';
-import { DropdownButton, MenuItem, Checkbox, Button, Collapse, ButtonToolbar } from 'react-bootstrap';
 import ReactTooltip from 'react-tooltip';
 
 
@@ -128,155 +127,6 @@ export default class WorkflowView extends DefaultItemView {
     }
 
 }
-
-
-export class WorkflowGraphSectionControls extends React.Component {
-
-    static analysisStepsSet(context){
-        if (!Array.isArray(context.steps)) return false;
-        if (context.steps.length === 0) return false;
-        return true;
-    }
-
-    static keyTitleMap = {
-        'detail'    : 'Analysis Steps',
-        'basic'     : 'Basic Inputs & Outputs',
-        'cwl'       : 'CWL Graph'
-    }
-
-    chartTypeDropdown(){
-        var detail = WorkflowGraphSectionControls.analysisStepsSet(this.props.context) ? (
-            <MenuItem eventKey='detail' active={this.props.showChartType === 'detail'}>
-                Analysis Steps
-            </MenuItem>
-        ) : null;
-    
-        var basic = (
-            <MenuItem eventKey='basic' active={this.props.showChartType === 'basic'}>
-                Basic Inputs & Outputs
-            </MenuItem>
-        );
-
-        return (
-            <DropdownButton id="detail-granularity-selector" key="chart-type"
-                className="for-state-showChart" pullRight
-                onSelect={this.props.onChangeShowChartType}
-                title={WorkflowGraphSectionControls.keyTitleMap[this.props.showChartType]}>
-                { basic }{ detail }
-            </DropdownButton>
-        );
-    }
-
-    rowSpacingTypeDropdown(){
-        if (typeof this.props.rowSpacingType !== 'string' || typeof this.props.onChangeRowSpacingType !== 'function') {
-            return null;
-        }
-        return (
-            <RowSpacingTypeDropdown currentKey={this.props.rowSpacingType} onSelect={this.props.onChangeRowSpacingType} key="row-spacing-type"/>
-        );
-    }
-
-    fullScreenButton(){
-        var { fullscreenViewEnabled, onToggleFullScreenView } = this.props;
-        if( typeof fullscreenViewEnabled === 'boolean' && typeof onToggleFullScreenView === 'function'){
-            return (
-                <Button onClick={onToggleFullScreenView} className="for-state-fullscreenViewEnabled"
-                    data-tip={!fullscreenViewEnabled ? 'Expand to full screen' : null} key="full-screen-btn">
-                    <i className={"icon icon-fw icon-" + (!fullscreenViewEnabled ? 'expand' : 'compress')}/>
-                </Button>
-            );
-        }
-        return null;
-    }
-
-    parametersCheckbox(){
-        if (typeof this.props.showParameters !== 'boolean' || typeof this.props.onToggleShowParameters !== 'function'){
-            return null;
-        }
-        return (
-            <Checkbox checked={this.props.showParameters} onChange={this.props.onToggleShowParameters}
-                className="checkbox-container for-state-showParameters" key="params-checkbox">
-                Show Parameters
-            </Checkbox>
-        );
-    }
-
-    referenceFilesCheckbox(){
-        if (typeof this.props.showReferenceFiles !== 'boolean' || typeof this.props.onToggleReferenceFiles !== 'function') return null;
-        return (
-            <Checkbox checked={this.props.showReferenceFiles} onChange={this.props.onToggleReferenceFiles} key="ref-files-checkbox"
-                disabled={this.props.isReferenceFilesCheckboxDisabled} className="checkbox-container for-state-showReferenceFiles">
-                Show Reference Files
-            </Checkbox>
-        );
-    }
-
-    /**
-     * @param {...JSX.Element} element - Element(s) to wrap in controls wrapper.
-     * @returns {JSX.Element} Workflow Controls Element.
-     */
-    wrapper(){
-        return (
-            <CollapsibleItemViewButtonToolbar
-                children={[...arguments]}
-                constantButtons={this.fullScreenButton()}
-                windowWidth={this.props.windowWidth} />
-        );
-    }
-
-    render(){
-        return this.wrapper(this.referenceFilesCheckbox(), this.parametersCheckbox(), this.chartTypeDropdown(), this.rowSpacingTypeDropdown());
-    }
-}
-
-
-export class RowSpacingTypeDropdown extends React.Component {
-    
-    static propTypes = {
-        'onSelect' : PropTypes.func.isRequired,
-        'currentKey' : Graph.propTypes.rowSpacingType,
-    }
-
-    static rowSpacingTypeTitleMap = {
-        'stacked' : 'Stack Nodes',
-        'compact' : 'Center Nodes',
-        'wide' : 'Spread Nodes'
-    }
-
-    render(){
-
-        var currentKey = this.props.currentKey;
-
-        var stacked = (
-            <MenuItem eventKey='stacked' active={currentKey === 'stacked'}>
-                { RowSpacingTypeDropdown.rowSpacingTypeTitleMap['stacked'] }
-            </MenuItem>
-        );
-    
-        var compact = (
-            <MenuItem eventKey='compact' active={currentKey === 'compact'}>
-                { RowSpacingTypeDropdown.rowSpacingTypeTitleMap['compact'] }
-            </MenuItem>
-        );
-
-        var spread = (
-            <MenuItem eventKey='wide' active={currentKey === 'wide'}>
-                { RowSpacingTypeDropdown.rowSpacingTypeTitleMap['wide'] }
-            </MenuItem>
-        );
-    
-    
-        return (
-            <DropdownButton id={this.props.id || "rowspacingtype-select"}
-                pullRight onSelect={this.props.onSelect} title={RowSpacingTypeDropdown.rowSpacingTypeTitleMap[currentKey]}>
-                { stacked }{ compact }{ spread }
-            </DropdownButton>
-        );
-    }
-
-}
-
-
 
 
 export class WorkflowGraphSection extends React.Component {
